[Release]
Added CMakeLists
Standardized utf8_tools to follow utfcpp style.
Proper crossplatform input/output utf-8 functions with neat throwables.
Totally replaced utf8_scanf(unstable and for tests only) with neater any-overflow-safe utf8_getstring function.
Crossplatform argv support  vector of utf-8 strings (wmain for windows with wchar_t is converted to utf-8 string)

+++ include/utf8_tools.h
@@ -8,9 +8,9 @@ std::string utf8_join(const std::string &, const std::string & = ", ");
 
 std::string utf8_reverse(const std::string &);
 
-std::string utf8_substr(const std::string &, int, int= 0);
+std::string utf8_substr(const std::string &, size_t, size_t = 0);
 
-std::string utf8_repeat(const std::string &, int);
+std::string utf8_repeat(const std::string &, size_t);
 
 std::vector<std::string> utf8_to_vector(const std::string &);
